refactor(Step5) : RacingGame constructor 수정 및 도메인 위치변경
- 생성자에 List<String> 으로 입력 하면, Car 객체를 생성해서 RacingGame 개최하도록 수정함.
- 주입용으로 만들었던 RandomLogic 들어간 RandomEngine클래스 이름 변경 하고 domain 패키지로 이동함. (DefaultRandomEngine)

issue #596
