+++ src/main/java/racingGame/RacingGamePlayer.java
@@ -1,17 +1,13 @@
 package racingGame;
 
 import racingGame.domain.Car;
-import racingGame.domain.Engine;
 import racingGame.domain.RacingGame;
 import racingGame.ui.InputView;
 import racingGame.ui.ResultView;
 
-import java.util.Arrays;
 import java.util.stream.Collectors;
 
 public class RacingGamePlayer {
-	private static final Engine randomEngine = new RandomEngine();
-
 	private InputView inputView;
 	private ResultView resultView;
 
@@ -22,10 +18,7 @@ public class RacingGamePlayer {
 
 	public void racingStart() {
 		inputView.startInput();
-		RacingGame racingGame = new RacingGame(
-				Arrays.stream(inputView.getRacers())
-						.map(racer -> new Car(racer, randomEngine))
-						.toArray(Car[]::new));
+		RacingGame racingGame = new RacingGame(inputView.getRacers());
 
 		resultView.printResultStart();
 		for (int i = 0; i < inputView.getRacingTime(); i++) {
